Refactor relevant pytest fixtures into 'conftest.py'
